requirements.txt
Flask
googlesearch-python
requests
beautifulsoup4